deploy: ce5e24f80310253fad9f4769e30519547b1da6fb

+++ include/roboteam_ai/control/RobotCommand.h
@@ -18,7 +18,7 @@ class RobotCommand {
     Vector2 pos = Vector2();
     Vector2 vel = Vector2();
     Angle angle = Angle();
-    float dribbler = 0;
+    int dribbler = 0;
     bool kicker = false;
     double kickerVel = 0;
     bool kickerForced = false;
